moving stack trace class out into a header for portability
